Separate Share and Invite components
