+++ frontend/src/components/common/index.ts
@@ -1,2 +1,4 @@
+export { default as Invite } from './Invite.vue';
+export { default as Share } from './Share.vue';
+export { default as ShareButton } from './ShareButton.vue';
 export { default as SyncButton } from './SyncButton.vue';
-export { default as InviteButton } from './InviteButton.vue';
