--- frontend/src/components/object/share/index.ts
@@ -1,2 +1,0 @@
-export { default as ShareLinkContent } from './ShareLinkContent.vue';
-export { default as ShareObjectButton } from './ShareObjectButton.vue';
